refactor code in both tls and http
